% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdm_IBD_v7_TF.R
\name{rdm_IBD}
\alias{rdm_IBD}
\title{Calculates Isolation by Distance (IBD) residuals for each pair of individuals in a population}
\usage{
rdm_IBD(Gen_raw, Geo_raw, Dist_method = 1, IBD_method = 1)
}
\arguments{
\item{Gen_raw}{A string specifying the path to a file of genotype data in GENEPOP format, with alleles specified
using three digits. The GENEPOP file must have the extension \emph{.gen}.}

\item{Geo_raw}{A string specifying the path to a text file showing the name (1st column) and geographical x and y 
coordinates (2nd and 3rd columns) of each individual (each row). The first row needs to specify the column headings, 
and all entries need to be tab-delimited.}

\item{Dist_method}{An integer from 1 to 6 specifying the method used to calculate genetic distance among individuals. 
Integers 1 to 6 correspond to the methods incorporated into the following six functions from the
\code{poppr} R package, respectively: \code{diss.dist}, \code{nei.dist}, \code{rogers.dist}, \code{reynolds.dist},
\code{edwards.dist}, and \code{provesti.dist}. The default value is 1.}

\item{IBD_method}{An integer specifying the method used to calculate the IBD residual for each pair of individuals. 
A value of 1 means that the residuals are calculated by fitting a straight line to all pairs of genetic and 
geographic distances, and measuring the vertical distance from each pair to the fitted line. The fitted straight line 
represents the combinations of genetic and geographic distances that are expected from IBD. A value of 2 means
that a non-linear curve of the form \emph{y = a + b*}(1\emph{-}exp(\emph{-}exp(\emph{c})\emph{*x})) is fitted instead of a straight 
line, where \emph{a}, \emph{b} and \emph{c} are the fitted parameters. The default value is 1. The function 
\code{rdm_IBD} automatically plots the data points together with the fitted line or curve.}
}
\value{
An object of class \code{dist}, containing a matrix showing the IBD residuals for pairs of individuals.
}
\description{
\code{rdm_IBD}: Reads genotype data for individuals in a population together with the geographic 
coordinates of the individuals, models overall IBD trend for the population, and then calculates IBD residuals for each pair of individuals.
}
\examples{
rdm_IBD("files/genotypes.gen", "files/coordinates.txt")
}
