% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdm_residual_v9_TF_TQ.R
\name{rdm_residual}
\alias{rdm_residual}
\title{Creates and plots IBD residuals for pairs of individuals in a population, represented as line segments over a 
 defined landscape}
\usage{
rdm_residual(IBD.res, Geo_raw, min.dist = 1, max.dist = Inf,
  n_resolution = 50, proj = sp::CRS("+init=epsg:4326"))
}
\arguments{
\item{IBD.res}{An object of class \code{dist}, containing a matrix with the IBD residuals for pairs of 
individuals in a population.}

\item{Geo_raw}{A string specifying the path to a text file showing the name (1st column) and geographical x and y 
coordinates (2nd and 3rd columns) of each individual (each row). The first row needs to specify the column headings, 
and all entries need to be tab-delimited.}

\item{min.dist}{IBD residuals are only calculated for pairs of individuals that are separated by a distance greater
than \code{min.dist}. The default value is 1.}

\item{max.dist}{IBD residuals are only calculated for pairs of individuals that are separated by a distance less
than \code{max.dist}. The default value is Inf.}

\item{n_resolution}{Specifies the number of cells that the landscape is divided into, for the purposes of creating
the 3-D plot. The landscape is divided into \code{n_resolution} cells along both coordinate axes. Individuals that
appear in the same cell are represented by a single point on the plot. The default value is 50.}

\item{proj}{The coordinate system that is used for plotting. The default is EPSG 4326.}
}
\value{
An object of class \code{SpatialLinesDataFrame}, containing coordinates of the line segments joining each 
 pair of individuals and the corresponding IBD residuals.
}
\description{
\code{rdm_residual}: Reads the IBD residuals for each pair of individuals in a population (as calculated 
 by function \code{rdm_IBD}) together with their geographic coordinates, and then creates a 
 3-D graph showing the residuals for each pair of individuals, represented as a line segment over the
 landscape considered. The 3-D graph can be rotated to obtain a better view.
}
\examples{
IBD_res <- rdm_IBD("files/genotypes.gen", "files/coordinates.txt")
rdm_residuals(IBD_res, "files/coordinates.txt")
}
