% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdm_resistance_v8_TF_TQ.R
\name{rdm_resistance}
\alias{rdm_resistance}
\title{Produces environmental resistance to dispersal over a landscape, for individuals in a population}
\usage{
rdm_resistance(IBD.res, Res_SLDF, nrows = 30, ncols = 30,
  conf_intervals = 0.95, random_rep = 1000,
  outputfile = "resistance_map.csv")
}
\arguments{
\item{IBD.res}{An object of class \code{dist}, containing a matrix with the IBD residuals for pairs of 
individuals in a population.}

\item{Res_SLDF}{An object of class \code{SpatialLinesDataFrame}, containing the IBD residuals for each pair of 
individuals in a population and the coordinates of line segments connecting the pairs of individuals over the 
defined landscape. This object is produced by the function \code{rdm_residual}.}

\item{nrows}{Number of grid cells in each row of the raster map of resistance. The default value is 30.}

\item{ncols}{Number of grid cells in each column of the raster map of resistance. The default value is 30.}

\item{conf_intervals}{The coverage of the confidence interval generated for each resistance value in a grid cell, 
expressed as a proportion. This interval measures the uncertainty in the (observed) resistance in each cell. 
If a cell has no intersecting line segments or only one, then a confidence interval cannot be calculated and the cell
is no longer considered for further calculation. If the confidence interval does not overlap 0, then the resistance 
is statistically different from 0. The default value is 0.95 (corresponding to 95\% intervals).}

\item{random_rep}{Number of random resamples of the IBD residuals used to construct the null distribution of 
resistances in each grid cell. For a grid cell with \emph{n} intersecting line segments and \emph{n} corresponding IBD residuals,
a random resample consists of randomly sampling \emph{n} IBD residuals from the set of all IBD residuals over the entire
landscape, without replacement. If the observed resistance is above a threshold percentile of the null
distribution, then the cell is inferred to be a genetic barrier. If instead the observed resistance is below 
another threshold percentile of the null distribution, then the cell is inferred to be a genetic corridor. The default 
value is 1,000.}

\item{outputfile}{Name of the .csv file that is produced, containing resistance information in the output data frame. 
The default is 'resistance_map.csv'.}
}
\value{
A data frame with resistance information for each grid cell in the landscape with more than one 
 intersecting line segment. Each row corresponds to one cell, and the eight different columns refer to the 
 resistance, the number of intersecting line segments, the x and y coordinates specifying the location of the mid-point of the cell, 
 the lower and upper limits of the confidence interval for the resistance, the sign of the product of the lower and upper 
 limits of the confidence interval, and the percentile of the null distribution of resistances corresponding 
 to the observed resistance. When calculating this data, the function \code{rdm_resistance} prints out messages
 showing the current stage of calculation. There are four stages: Calculating (1) the resistances, (2) the
 numbers of intersecting line segments, (3) the lower limits of the confidence intervals for the resistances,
 and (4) the upper limits of the confidence intervals for the resistances.
}
\description{
\code{rdm_resistance}: Reads in IBD residuals for pairs of individuals in a population, represented
 as line segments connecting the pairs of individuals over a defined landscape, and then applies a novel method 
 (Tang et al., in review) to the residuals to calculate environmental resistance to dispersal over the landscape. 
 Essentially, the method splits the landscape into a grid of cells and for each cell, calculates the 
 resistance as the average of the IBD residuals of all line segments that intersect the cell.
}
\examples{
IBD_res <- rdm_IBD("files/genotypes.gen", "files/coordinates.txt")
Res_SLDF <- rdm_residuals(IBD_res, "files/coordinates.txt")
rdm_resistance(IBD_res, Res_SLDF)
}
\references{
Tang, Q., Fung, T., Rheindt, F.E. (In review). ResDisMapper: An R package for fine-scale mapping of environmental resistance to dispersal.
}
